[REF] web,*: move templates out of assets.js
This commit moves the `templates` XML document containing all static
templates out of the `@web/core/assets.js` file to move it to its own
file.

This has been done in order to separate concerns: loading assets
(`assets.js`) and registering/accessing static templates
(`templates.js`).
