+++ addons/hr_attendance/static/src/public_kiosk/public_kiosk_app.js
@@ -4,7 +4,7 @@ import {App, whenReady, Component, useState} from "@odoo/owl";
 import { CardLayout } from "@hr_attendance/components/card_layout/card_layout";
 import { KioskManualSelection } from "@hr_attendance/components/manual_selection/manual_selection";
 import { makeEnv, startServices } from "@web/env";
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { _t } from "@web/core/l10n/translation";
 import { MainComponentsContainer } from "@web/core/main_components_container";
 import { useService, useBus } from "@web/core/utils/hooks";
