+++ addons/mail/static/src/discuss/core/public/boot.js
@@ -4,7 +4,7 @@ import { DiscussPublic } from "@mail/discuss/core/public/discuss_public";
 
 import { mount, whenReady } from "@odoo/owl";
 
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { MainComponentsContainer } from "@web/core/main_components_container";
 import { registry } from "@web/core/registry";
 import { makeEnv, startServices } from "@web/env";
