+++ addons/pos_epson_printer/static/src/app/epson_printer.js
@@ -2,7 +2,7 @@
 
 import { BasePrinter } from "@point_of_sale/app/printer/base_printer";
 import { _t } from "@web/core/l10n/translation";
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { createElement, append, createTextNode } from "@web/core/utils/xml";
 
 function ePOSPrint(children) {
