+++ addons/spreadsheet/__manifest__.py
@@ -55,6 +55,7 @@
             'web/static/src/core/browser/feature_detection.js',
             'web/static/src/core/registry.js',
             'web/static/src/core/assets.js',
+            'web/static/src/core/templates.js',
             'web/static/src/session.js',
             'web/static/src/env.js',
             'web/static/src/core/network/http_service.js',
