+++ addons/spreadsheet/static/src/public_readonly_app/main.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 import { App, whenReady } from "@odoo/owl";
 import { PublicReadonlySpreadsheet } from "./public_readonly";
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { makeEnv, startServices } from "@web/env";
 import { session } from "@web/session";
 import { _t } from "@web/core/l10n/translation";
