+++ addons/web/static/src/core/utils/render.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
 import { App, blockDom, Component, markup } from "@odoo/owl";
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { _t } from "@web/core/l10n/translation";
 
 export function renderToElement(template, context = {}) {
