+++ addons/web/static/src/legacy/js/public/public_root.js
@@ -8,7 +8,8 @@ import { registry } from '@web/core/registry';
 import lazyloader from "@web/legacy/js/public/lazyloader";
 
 import { makeEnv, startServices } from "@web/env";
-import { loadJS, templates } from '@web/core/assets';
+import { loadJS } from '@web/core/assets';
+import { templates } from '@web/core/templates';
 import { MainComponentsContainer } from "@web/core/main_components_container";
 import { browser } from '@web/core/browser/browser';
 import { _t } from "@web/core/l10n/translation";
