+++ addons/web/static/src/start.js
@@ -3,7 +3,7 @@
 import { makeEnv, startServices } from "./env";
 import { localization } from "@web/core/l10n/localization";
 import { session } from "@web/session";
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { hasTouch } from "@web/core/browser/feature_detection";
 import { _t } from "@web/core/l10n/translation";
 import { App, Component, whenReady } from "@odoo/owl";
