+++ addons/web/static/tests/core/tooltip/tooltip_service_tests.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
 import { App, Component, useState, xml } from "@odoo/owl";
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { browser } from "@web/core/browser/browser";
 import { hotkeyService } from "@web/core/hotkeys/hotkey_service";
 import { MainComponentsContainer } from "@web/core/main_components_container";
