+++ addons/web/static/tests/helpers/utils.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { templates } from "@web/core/assets";
+import { templates } from "@web/core/templates";
 import { browser } from "@web/core/browser/browser";
 import { isMacOS } from "@web/core/browser/feature_detection";
 import { download } from "@web/core/network/download";
